refactor: Simplify imports in CommentProcessorFactory and move SectionUtil to core package
