+++ engine/src/main/java/pro/verron/officestamper/core/SectionUtil.java
@@ -1,4 +1,4 @@
-package org.wickedsource.docxstamper.util;
+package pro.verron.officestamper.core;
 
 import org.docx4j.XmlUtils;
 import org.docx4j.jaxb.Context;
