+++ engine/src/main/java/pro/verron/officestamper/preset/CommentProcessorFactory.java
@@ -9,13 +9,13 @@ import org.docx4j.openpackaging.packages.WordprocessingMLPackage;
 import org.docx4j.wml.*;
 import org.jvnet.jaxb2_commons.ppp.Child;
 import org.springframework.lang.Nullable;
-import org.wickedsource.docxstamper.util.*;
+import org.wickedsource.docxstamper.util.DocumentUtil;
+import org.wickedsource.docxstamper.util.ObjectDeleter;
+import org.wickedsource.docxstamper.util.ParagraphUtil;
+import org.wickedsource.docxstamper.util.RunUtil;
 import org.wickedsource.docxstamper.util.walk.BaseDocumentWalker;
 import pro.verron.officestamper.api.*;
-import pro.verron.officestamper.core.CommentUtil;
-import pro.verron.officestamper.core.DocxStamper;
-import pro.verron.officestamper.core.PlaceholderReplacer;
-import pro.verron.officestamper.core.StandardParagraph;
+import pro.verron.officestamper.core.*;
 
 import java.io.IOException;
 import java.io.OutputStream;
